requirements.txt
hydra-core >= 1.3.2
opencv-python >= 4.5.5.62
scipy >= 1.3.0
scikit-image >= 0.13.0
av >= 10.0.0
six >= 1.16.0 
mediapipe==0.10.20
torch
torchvision
torchaudio
ollama
pydantic
pynput
